Unify more classification test helpers
